fixed all errors and cleaned everything up - will now return to solar plant error management

+++ local/include/microsun/m_customer.h
@@ -53,7 +53,7 @@ class Customer : public Php::Base
                 }
             Php::Value phpgetID()
                 {
-                return (double) this->ID;
+                return (int) this->ID;
                 }
 
             Php::Value phpgetCompany()
